requirements.txt
fastapi>=0.120.0
uvicorn[standard]>=0.38.0
sqlalchemy[asyncio]>=2.0.19
psycopg[binary]>=3.0.0  # asyncio-compatible PostgreSQL driver
python-dotenv>=1.0.0
tiktoken>=0.4.0
pydantic>=1.10.12  # consider pydantic v2 migration carefully
python-multipart>=0.0.6
PyPDF2>=3.0.1

# ✅ Modern LangChain ecosystem
langchain==1.0.2
langchain-core==1.0.1
langchain-openai>=0.1.0
langchain-postgres>=0.0.9
langchain-community>=0.0.29

# ✅ OpenAI client
openai>=1.3.0

# ✅ For pgvector compatibility
pgvector>=0.3.0
